pyproject.toml
[tool.poetry]
name = "agproject"
version = "0.1.0"
description = ""
authors = ["Adam <adamrek.st9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "2.0"
matplotlib = "^3.10.0"
pillow = "^11.0.0"
rich = "^13.9.4"
numba = "^0.60.0"
scikit-image = "^0.25.0"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
